# public/swagger-spec.json
{"openapi":"3.0.0","paths":{"/users/registration":{"post":{"operationId":"UserController_registration","summary":"Sing up User in to application","description":"## Регистрация нового пользователя\n Для регистрации требуется указать следующие поля:\n\n ```json\n {\n  \"email\": \"string\",\n  \"password\": \"string\" \n}\n ```\n","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/login":{"post":{"operationId":"UserController_login","summary":"Sing in User in application","description":"## Авторизация в приложении\n\nПришлите логин и пароль аккаунта и получите `accessToken` и `refreshToken`. \n* **Условия**:\nПри неверном введении пароля 5 раз повторный ввод пароля можно будет сделать через 1 час.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/verify":{"get":{"operationId":"UserController_verificationUser","summary":"User verification using a link to email","description":"## Проверка подлинности\nНа указанный адрес регистратора отправлено письмо. Требуется подтверждение.\nПройдите по ссылке указанный в почтовом ящике или пришлите `secretKey` ключ для подтвержения почтового ящика.\n","parameters":[{"name":"secretKey","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/users/password/forgot":{"patch":{"operationId":"UserController_forgotPassword","summary":"Step 1: Forgotten password recovery","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserForgotPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["User"]}},"/users/password/refresh/verify":{"patch":{"operationId":"UserController_refreshVerify","summary":"Refresh request link for reset password","description":"## Повторное получение ссылки на восстановление пароля","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserForgotPasswordDto"}}}},"responses":{"200":{"description":""}},"tags":["User"]}},"/users/password/forgot/verify":{"patch":{"operationId":"UserController_forgotVerify","summary":"Step 2: User verification using a link to email. ","description":"","parameters":[{"name":"verification","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/users/password/reset":{"patch":{"operationId":"UserController_resetPassword","summary":"Step 3: User reset password. ","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResetPasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResetPasswordDto"}}}}},"tags":["User"]}},"/users/change/password":{"patch":{"operationId":"UserController_changePassword","summary":"Change password","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserChangePasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserChangePasswordDto"}}}}},"tags":["User"],"security":[{"bearer":[]}]}},"/users/create":{"post":{"operationId":"UserController_createUser","summary":"User creation","description":"## Creates a new user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"tags":["User"],"security":[{"bearer":[]}]}},"/users":{"get":{"operationId":"UserController_findAllUsers","summary":"List of all users","description":"## Лист всех пользователей","parameters":[],"responses":{"200":{"description":""}},"tags":["User"],"security":[{"bearer":[]}]}},"/users/{id}":{"delete":{"operationId":"UserController_archiveUser","summary":"User archiving","description":"## Архивация пользоватлей","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"active","required":true,"in":"query","schema":{"enum":["true","false"],"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"],"security":[{"bearer":[]}]}},"/roles/create":{"post":{"operationId":"RolesController_create","summary":"Creating role for user","description":"## Role create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RolesDto"}}}},"responses":{"201":{"description":""}},"tags":["Roles"]}},"/roles/{id}":{"patch":{"operationId":"RolesController_update","summary":"Updating role","description":"## Изменение данных роли","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRolesDto"}}}},"responses":{"200":{"description":""}},"tags":["Roles"]}},"/roles/list":{"get":{"operationId":"RolesController_findAllRoles","summary":"Get list of all roles","description":"## Список всех ролей","parameters":[],"responses":{"200":{"description":""}},"tags":["Roles"],"security":[{"bearer":[]}]}},"/profile":{"post":{"operationId":"ProfileController_createPersona","summary":"Создание профиля пользователя","description":"## Создание профиля","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}}},"tags":["Profile"]}},"/profile/me/update":{"patch":{"operationId":"ProfileController_updatePersona","summary":"Фото или аватар пользователя","description":"## Создание профиля","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}}},"tags":["Profile"]}},"/profile/me/update/location":{"patch":{"operationId":"ProfileController_insertOrUpdateAddressPersona","summary":"Данные о месте проживания","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}}},"tags":["Profile"]}},"/profile/me/update/avatar":{"patch":{"operationId":"ProfileController_updateAvatarPersona","summary":"Фото или аватар пользователя","description":"## Создание профиля","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}}},"tags":["Profile"]}},"/profile/me/update/socials":{"patch":{"operationId":"ProfileController_updateSocialPersona","summary":"Социальные сети пользователя","description":"## Создание профиля","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}}},"tags":["Profile"]}},"/profile/me":{"get":{"operationId":"ProfileController_myDataPersona","summary":"Данные пользователя","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePersonaDto"}}}}},"tags":["Profile"]}}},"info":{"title":"Cleverdeus microservices","description":"## App Documentation\n\nRead Here v1","version":"1.0.0","contact":{}},"tags":[{"name":"REST API","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserDto":{"type":"object","properties":{"email":{"type":"string","example":"test@cleverdeus.com","format":"email","uniqueItems":true,"minLength":5,"maxLength":255},"password":{"type":"string","example":"password","format":"string","minLength":5,"maxLength":16}},"required":["email","password"]},"UserForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"test@cleverdeus.com","format":"email","uniqueItems":true,"minLength":5,"maxLength":255}},"required":["email"]},"UserResetPasswordDto":{"type":"object","properties":{"password_new":{"type":"string","format":"string","minLength":5,"maxLength":16},"password_confirm":{"type":"string","format":"string","minLength":5,"maxLength":16},"verificationKey":{"type":"string"}},"required":["password_new","password_confirm","verificationKey"]},"UserChangePasswordDto":{"type":"object","properties":{"password":{"type":"string","format":"string","minLength":5,"maxLength":16},"password_new":{"type":"string","format":"string","minLength":5,"maxLength":16},"password_confirm":{"type":"string","format":"string","minLength":5,"maxLength":16}},"required":["password","password_new","password_confirm"]},"RolesDto":{"type":"object","properties":{"name":{"type":"string","example":"Guest"},"permissions":{"type":"string"}},"required":["name","permissions"]},"UpdateRolesDto":{"type":"object","properties":{"permissions":{"type":"string"},"priority":{"type":"number"}},"required":["permissions","priority"]},"Map":{"type":"object","properties":{}},"ProfilePersonaDto":{"type":"object","properties":{"title":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"maidenName":{"type":"string"},"nickName":{"type":"string"},"about":{"type":"string"},"gender":{"type":"string"},"relationship":{"type":"string"},"language":{"type":"string"},"address":{"$ref":"#/components/schemas/Map"},"birthDate":{"format":"date-time","type":"string"},"speakLanguage":{"type":"array","items":{"type":"string"}},"socials":{"$ref":"#/components/schemas/Map"}},"required":["title","firstName","lastName","maidenName","nickName","about","gender","relationship","language","address","birthDate","speakLanguage","socials"]}}}}